requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
langchain==0.2.16
langchain-core==0.2.41
langchain-community==0.2.16
langchain-google-genai==1.0.10
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-asyncio==0.23.8
httpx==0.27.0
